Add bill run entity and use it in workflow bill run scenarios
Consolidates the charge version + bill run boilerplate duplicated across scenarios into a shared entity, defaulting the batch type to annual and letting callers override it.
