+++ tests/support/scenarios/licence-with-workflow-and-two-part-tariff-bill-run.scenario.js
@@ -1,8 +1,7 @@
-import billRunData from '../data/bill-run.data.js'
 import workflowData from '../data/workflow.data.js'
-import buildChargeVersionEntity from '../entities/charge-version.entity.js'
+import buildBillRunEntity from '../entities/bill-run.entity.js'
 import buildLicenceEntity from '../entities/licence.entity.js'
-import { today, yesterday } from '../helpers/date.helpers.js'
+import { yesterday } from '../helpers/date.helpers.js'
 
 export const title = 'Licence in workflow, and a two-part tariff bill run'
 export const description =
@@ -14,26 +13,16 @@ export const description =
  * This is omitted from the scenario name and description to keep them concise, but is still part of the scenario.
  */
 export default function (calculatedDates) {
-  const licenceEntity = buildLicenceEntity()
-  const chargeVersionEntity = buildChargeVersionEntity(
-    licenceEntity.company,
-    licenceEntity.address,
-    licenceEntity.licence,
-    licenceEntity.licenceVersionPurpose
-  )
-
   const {
     billingPeriods: {
       twoPartTariff: [twoPartTariffDates]
     }
   } = calculatedDates
 
-  const billRun = billRunData()
+  const licenceEntity = buildLicenceEntity()
+  const billRunEntity = buildBillRunEntity(licenceEntity, twoPartTariffDates)
 
-  billRun.createdAt = today()
-  billRun.batchType = 'two_part_tariff'
-  billRun.fromFinancialYearEnding = new Date(twoPartTariffDates.endDate).getUTCFullYear()
-  billRun.toFinancialYearEnding = new Date(twoPartTariffDates.endDate).getUTCFullYear()
+  billRunEntity.billRun.batchType = 'two_part_tariff'
 
   const workflow = workflowData(licenceEntity.licence)
 
@@ -44,8 +33,7 @@ export default function (calculatedDates) {
 
   return {
     ...licenceEntity,
-    ...chargeVersionEntity,
-    billRun,
+    ...billRunEntity,
     workflow
   }
 }
